Updated README.md
This app can be added upon in many ways. 
This is a combination of two generators and the future development of this app might be improved by having the background color gradient generated be in the Sweet Tweet to be shared on Twitter as animation or other content as intended.